refactor: view, viewModel 파일 분리
